requirements.txt
asgiref==3.8.1
astroid==3.2.4
dill==0.3.8
dj-database-url==2.2.0
Django==5.1.1
django-tastypie==0.14.7
isort==5.13.2
mccabe==0.7.0
platformdirs==4.3.6
psycopg2-binary==2.9.9
pylint==3.2.7
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dateutil==2.9.0.post0
python-mimeparse==2.0.0
pytz==2024.2
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
tomlkit==0.13.2
typing_extensions==4.12.2
